requirements.txt
confluent-kafka
fastapi
uvicorn
python-socketio
geopy
folium
python-multipart

